// lib/types.ts
export interface TipoCliente {
  id: number;
  nombreTipo: string;
  descripcion?: string;
  estado: "Activo" | "Inactivo";
}

export interface SectorIndustria {
  id: number;
  nombreSector: string;
  descripcion?: string;
  estado: "Activo" | "Inactivo";
}

export interface TipoRelacionComercial {
  id: number;
  nombre: string;
  descripcion?: string;
}

export interface CondicionPago {
  id: number;
  nombreCondicion: string;
  descripcion?: string;
}

export interface DocumentoClienteDefinicion {
  id: number;
  nombreDocumento: string;
  aplicaA: "Empresa" | "Persona" | "Institución" | "Todos";
  esObligatorio: boolean;
  tieneVencimiento: boolean;
}